+++ changelog.md
@@ -25,6 +25,7 @@ fix:
 - fixed simoultaneous log and density th2 pyplots
 - fixed verbosity option
 - wrong indentation caused missing cpp imports
+- common flows were randomly created due to the usage of dict instead of orderedDict
 
 cmgrdf commit bump
 - Plots cleaned, just TH object saved in TFiles